+++ apps/admin/lib/auth/auth-bff.ts
@@ -492,129 +492,3 @@ export async function zitadelTotp(
   const outcome = parseLoginResponse(body);
   return { ...outcome, setCookies: readAllSetCookies(res) };
 }
-
-// ---------------------------------------------------------------------------
-// Google sign-in through Zitadel's IDP-intent flow (#524 phase 3c-2), the
-// merchant/admin pair of the endpoints apps/storefront/lib/auth/
-// auth-bff-customer.ts drives for shoppers. Grouped separately from
-// zitadelLogin/zitadelTotp above so a later provider (Apple) has an
-// obvious place to land alongside these.
-
-/**
- * startZitadelIDPIntent submits {return_url} to
- * POST /auth/zitadel/idp/start and returns the authUrl the browser must be
- * redirected to. returnUrl must already be this app's own /auth/idp/finish
- * route, pre-validated server-side against auth-bff's ADMIN return-url
- * allowlist (a separate, narrower list than the storefront one — see
- * idpintent.go's StartIDPIntent doc and newZitadelHandlers in auth-bff's
- * cmd/server/main.go for why the two must never be swapped) — Zitadel does
- * not validate successUrl at all, so a bad value here is an open redirect
- * waiting to happen.
- */
-export async function startZitadelIDPIntent(returnUrl: string): Promise<string> {
-  const res = await fetch(`${base}/auth/zitadel/idp/start`, {
-    method: "POST",
-    headers: { "Content-Type": "application/json", ...internalAuthHeader() },
-    body: JSON.stringify({ return_url: returnUrl }),
-    cache: "no-store",
-  });
-
-  if (!res.ok) {
-    let body: { error?: string; message?: string } = {};
-    try {
-      body = await res.json();
-    } catch {
-      // ignore
-    }
-    throw new AuthBffError(
-      res.status,
-      body.error ?? "auth_bff_error",
-      body.message ?? `HTTP ${res.status}`,
-    );
-  }
-
-  const body = (await res.json()) as { auth_url?: string };
-  if (!body.auth_url) {
-    throw new AuthBffError(res.status, "unrecognised_response_shape", "start idp intent: no auth_url");
-  }
-  return body.auth_url;
-}
-
-interface ZitadelIDPFinishRequest {
-  authRequestId: string;
-  intentId: string;
-  intentToken: string;
-  workspaceTenant: string;
-  /** See ZitadelLoginRequest.userAgent above — the same fingerprinting and
-   *  email-OTP rate-limiting reasons apply identically to a Google
-   *  sign-in as to a password one. */
-  userAgent?: string;
-  forwardedFor?: string;
-}
-
-/**
- * zitadelIdpFinish submits {auth_request_id, intent_id, intent_token,
- * workspace_tenant} to POST /auth/zitadel/idp/finish — the merchant
- * counterpart of the storefront's finishCustomerIDPIntent, except this
- * endpoint runs the full merchant gauntlet (FGA membership, the MFA gate,
- * deviceguard, the email-OTP step-up) and mints `m8_session` on success,
- * exactly like zitadelLogin/zitadelTotp above. The 2xx body is handed to
- * parseLoginResponse rather than parsed here, so this function's outcome
- * union and cookie-forwarding behaviour are identical to zitadelLogin's.
- *
- * Deliberately takes ONLY intentId/intentToken for identity — never a
- * `user` value. Zitadel's redirect back to the browser can carry a `user`
- * query param, but it rides in a URL the browser followed and is
- * attacker-controlled; the caller (app/auth/idp/finish/route.ts) must
- * never read it for anything, and this function has no parameter for it
- * at all so there is nothing to accidentally forward. auth-bff's own
- * idpFinishRequest.User field doc makes the same point server-side.
- *
- * A non-2xx response's `error` code is preserved on the thrown
- * AuthBffError exactly as-is (no_admin_account, unexpected_idp,
- * email_not_verified, email_ambiguous, invalid_intent,
- * zitadel_unavailable, ...) so the caller can render a truthful, distinct
- * message for each rather than a single collapsed failure — there is no
- * account-enumeration concern here the way there is on the password path
- * (an intent id/token pair is not a guessable credential).
- */
-export async function zitadelIdpFinish(
-  req: ZitadelIDPFinishRequest,
-): Promise<LoginOutcome & { setCookies: string[] }> {
-  const headers: Record<string, string> = {
-    "Content-Type": "application/json",
-    ...internalAuthHeader(),
-  };
-  if (req.userAgent) headers["User-Agent"] = req.userAgent;
-  if (req.forwardedFor) headers["X-Forwarded-For"] = req.forwardedFor;
-
-  const res = await fetch(`${base}/auth/zitadel/idp/finish`, {
-    method: "POST",
-    headers,
-    body: JSON.stringify({
-      auth_request_id: req.authRequestId,
-      intent_id: req.intentId,
-      intent_token: req.intentToken,
-      workspace_tenant: req.workspaceTenant,
-    }),
-    cache: "no-store",
-  });
-
-  if (!res.ok) {
-    let body: { error?: string; message?: string } = {};
-    try {
-      body = await res.json();
-    } catch {
-      // ignore
-    }
-    throw new AuthBffError(
-      res.status,
-      body.error ?? "auth_bff_error",
-      body.message ?? `HTTP ${res.status}`,
-    );
-  }
-
-  const body = await res.json();
-  const outcome = parseLoginResponse(body);
-  return { ...outcome, setCookies: readAllSetCookies(res) };
-}
